auto-improve: audit-refactor 2.1: create cai_lib/audit/ package and migrate cost helpers

Refs damien-robotsix/robotsix-cai#891
