+++ cai_lib/audit/__init__.py
@@ -1,0 +1,1 @@
+"""cai_lib.audit — audit-side infrastructure (cost aggregation, etc.)."""
