+++ cai_lib/audit/modules.py
@@ -1,0 +1,1 @@
+"""Audit module loader — stub for future audit infrastructure."""
