+++ cai_lib/cmd_misc.py
@@ -12,7 +12,8 @@ from datetime import datetime, timedelta, timezone
 from pathlib import Path
 
 from cai_lib.config import *  # noqa: F403,F401
-from cai_lib.logging_utils import log_run, _load_cost_log, _load_outcome_counts, _row_ts
+from cai_lib.logging_utils import log_run
+from cai_lib.audit.cost import _load_cost_log, _load_outcome_counts, _row_ts
 from cai_lib.subprocess_utils import _run, _run_claude_p
 from cai_lib.github import (
     _gh_json, _set_labels, _transcript_dir_is_empty,
